+++ crates/dynamics-joint/src/continuous.rs
@@ -132,6 +132,10 @@ impl JointModel for JointModelContinuous {
         );
         v[0] * SpatialMotion::from_rotational_axis(&self.axis)
     }
+
+    fn bias(&self) -> SpatialMotion {
+        SpatialMotion::zero()
+    }
 }
 
 /// Data structure containing the mutable properties of a continuous joint.